put all the command stuff in run.go
